fix flaky 2nd loop run of capybara specs by changing to shared_examples from different files
